Fix(DemoGame) - Fix compile error + Replace PhaseInfo with transition_data.h

+++ include/game/components/bullethell/bullet_clearer.h
@@ -7,7 +7,7 @@ namespace Game::BulletHell
     {
         float max_size;
         int speed;
-        int lifetime = UNASSIGNED;
+        int lifetime = static_cast<int>(UNASSIGNED);
         BulletClearer() : max_size(10), speed(1) {}
         BulletClearer(const float size, const int speed) : max_size(size), speed(speed) {}
     };
